requirements.txt
Django==4.2.1
mistune==3.0.0rc1
pygments==2.13.0
lxml==4.9.2
jinja2==3.1.2
django-environ